pyproject.toml
[project]
name = "checkmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[all]>=0.124.0",
    "kafka-python>=2.3.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
